refactor: move integration and unit tests to certain folder
