+++ package.json
@@ -47,9 +47,11 @@
     "watch": "pnpm run clean & pnpm run watch:types & pnpm run watch:tsup",
     "test": "pnpm test:typecheck && pnpm test:unit",
     "test:typecheck": "tsc -p ./test",
-    "test:unit": "vitest run --test-timeout=60000",
+    "test:unit": "vitest run test/unit --test-timeout=60000",
+    "test:integration": "vitest run test/integration --test-timeout=60000",
     "test:watch": "vitest",
-    "test:coverage": "vitest run --coverage",
+    "test:coverage-unit": "vitest run test/unit --coverage",
+    "test:coverage-integration": "vitest run test/integration --coverage",
     "format:fix": "prettier --ignore-path 'dist/* docs/*' --write '**/*.{ts,json,md}'",
     "lint:fix": "eslint . --ignore-pattern dist --ext .ts --fix",
     "prepare": "husky install",
