+++ test/integration/multiSig.spec.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect } from 'vitest';
 import { Transaction } from '@mysten/sui/transactions';
 import { SuiClient, getFullnodeUrl } from '@mysten/sui/client';
-import { SuiAccountManager } from '../src/libs/suiAccountManager';
-import { MultiSigClient } from '../src/libs/multiSig';
+import { SuiAccountManager } from '../../src/libs/suiAccountManager';
+import { MultiSigClient } from '../../src/libs/multiSig';
 
 const ENABLE_LOG = false;
