+++ test/unit/libs/suiModel/suiOwnedObject.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { SuiOwnedObject } from '../../../src/libs/suiModel/suiOwnedObject';
+import { SuiOwnedObject } from '../../../../src/libs/suiModel/suiOwnedObject';
 
 describe('SuiOwnedObject', () => {
   it('should initialize with objectId', () => {
